requirements.txt
fastapi
uvicorn
openai
faiss-cpu
pypdf
tiktoken
python-dotenv
streamlit
requests
